style: apply cargo clippy --fix (byte-identical)
Auto-applied clippy suggestions across 23 files (87 -> 41 warnings; the
remaining are non-auto-fixable doc-formatting lints and justified #[allow]s).
All changes are semantics-preserving: needless-return, collapsible let-chains
(if x { if let .. } -> if x && let ..), redundant casts/closures, % == 0 ->
is_multiple_of, Builder::default() -> Builder (unit struct), doc-comment
markers. No clamp() (NaN-behavior) lint was applied. Verified byte-identical:
chip/lcWGS/multi-chr imputation md5 all unchanged vs baseline; 164 lib tests pass.

+++ src/haploid/pbwt.rs
@@ -749,7 +749,7 @@ fn phase_subwindow(
                                 if let Some(ref iv) = inv {
                                     // Local copying consensus: majority over the K nearest
                                     // PBWT neighbours (expand outward, skipping missing).
-                                    let p = iv[hh as usize];
+                                    let p = iv[hh];
                                     let (mut n0, mut n1) = (0i32, 0i32);
                                     let mut off = 1i32;
                                     while (n0 + n1) < nocall_k as i32 && off < n_haps as i32 {
